feat(db): PostgreSQL'den MSSQL'e geçiş + Korgün setup script
## Kapsam

PostgreSQL initdb 0xC0000409 (STATUS_STACK_BUFFER_OVERFLOW) sorunu
cikis yolu: DijiCatalog artık MSSQL kullanıyor (Korgun ERP ile aynı
instance uzerinde ayri DB). Bu sayede:
- Lokal PostgreSQL kurulum derdi bitiyor
- Korgun ERP tablolarina (StokKart/CariKart) ayni MSSQL uzerinden
  hizli erisim (LAN icinde, tunnel gerekmez)
- Drizzle MSSQL dialect, mssql node paketi

## Schema degisiklikleri

11 schema dosyasi MSSQL'e cevrildi (pgTable -> mssqlTable):
- tenants, users, categories, products, productImages, customers,
  catalogs, catalogItems, catalogCustomers, catalogFieldConfig
- uuid -> uniqueidentifier (default newid())
- text -> nvarchar(max) / varchar(N)
- timestamp -> datetime2 (default getdate())
- numeric -> decimal
- jsonb -> nvarchar(max) (JSON string olarak)
- Enum'lar -> varchar(20) + app-level validation (MSSQL'de enum yok)
- index/uniqueIndex/AnyMsSqlColumn

PG-eski semalar .trash-2026-09-15-schema-pg/ icinde yedeklendi
(geri donmek icin). .gitignore pattern'i zaten calisiyor.

## Baglanti katmani

- src/config/database.ts: pg.Pool -> mssql.ConnectionPool
  Drizzle node-mssql adapter. DATABASE_URL format mssql://user:pass@host[:port|/INSTANCE]/db?encrypt=...&trustServerCertificate=...
- src/config/env.ts: URL validation mssql:// veya sqlserver:// kabul ediyor
- src/db/helpers.ts: AnyPgColumn -> AnyMsSqlColumn

## Service duzeltmeleri

PG-spesifik SQL kullanan 6 yeri MSSEL-uyumlu hale getirdik:
- 'count(*)::int' -> 'count(*)' (3 dosya, 6 yer)
- ILIKE -> LOWER(col) LIKE LOWER(...) (catalog.service.ts, 3 yer)
  DB-agnostic yaklasim: MSSEL + PG + MySQL collation bagimsiz calisir

## Drizzle config

drizzle.config.ts: dialect 'postgresql' -> 'mssql'

## Package.json

- pg kaldirildi (mssql zaten var, ERP adapter da kullanir)
- @types/pg kaldirildi
- description guncellendi: PostgreSQL -> MSSQL

## Seed

scripts/seed-demo.ts: pg.Pool -> mssql.ConnectionPool.
DATABASE_URL parser eklendi, mssql:// format kabul ediyor.

## Yeni scriptler

scripts/setup-korgun-mssql.ps1 (~260 satır):
  Korgun makinesinde (Administrator) tek-komut kurulum:
  - MSSQL DB olusturma (create-mssql-db.js ile)
  - Repo klonla
  - npm install + build
  - .env olusturma (DATABASE_URL otomatik)
  - drizzle-kit push (schema olustur)
  - Seed (demo data)
  - NSSM service kurulumu (Windows service)
  - Firewall (port 3000)

server/scripts/create-mssql-db.js (~95 satır):
  mssql node paketi ile master DB'ye baglanip CREATE DATABASE
  veya IF NOT EXISTS kontrolu yapar. CLI args + env destekler.

## Kullanım (Korgun'de RDP'de PowerShell Admin)

.\setup-korgun-mssql.ps1 \
  -MssqlServer 'localhost' \
  -MssqlPort 49746 \
  -MssqlInstance 'ABKA' \
  -MssqlUser 'sa' \
  -MssqlPassword 'Passw0rd!' \
  -DbName 'DijiCatalog' \
  -AppDir 'C:\\digicatalog' \
  -ServerIP '192.168.2.67'

Stats: 23 files changed, 775 insertions(+), 346 deletions(-)

+++ server/src/services/customer.service.ts
@@ -104,7 +104,7 @@ export const listCustomers = async (
   const orderFn = sortOrder === 'asc' ? asc : desc;
 
   const [{ count }] = await db
-    .select({ count: sql<number>`count(*)::int` })
+    .select({ count: sql<number>`count(*)` })
     .from(customers)
     .where(tenantAnd(customers, tenantId, ...conditions));
